privacy-center: Extract Consents types and fix spelling
